refactor(web): split events handler helpers [OPE-395]
