+++ crates/opengoose-web/src/handlers/events/mod.rs
@@ -1,0 +1,9 @@
+mod poll;
+mod snapshot;
+mod stream;
+
+pub use poll::list_event_history;
+pub use stream::stream_events;
+
+#[cfg(test)]
+mod tests;
